# packages/types/src/break.ts
/**
 * Break plan type - extensible enum
 */
export type BreakPlanType = 'breath' | 'walk' | string;

/**
 * Location with address and coordinates (for full plan object)
 */
export interface Location {
  address: string;
  coordinates: {
    lat: number;
    lng: number;
  };
}

/**
 * Simple location with only coordinates (for request)
 */
export interface LocationCoordinates {
  lat: number;
  lng: number;
}

/**
 * Full break plan object
 */
export interface BreakPlan {
  id: string; // UUID
  title: string;
  description: string;
  duration: number; // minutes
  location: Location;
  type: BreakPlanType;
  feeling: string; // Original user feeling/state when plan was requested
  chosen: boolean; // True when user selects this plan
  order: number; // Order in which plans appear in frontend (0, 1, 2, ...)
  created_at: string; // ISO 8601
  updated_at: string; // ISO 8601
  user_id: string;
}

/**
 * Request body for creating a break plan
 * Server auto-generates id and created_at
 */
export interface CreateBreakPlanRequest {
  duration: number; // minutes
  location: LocationCoordinates; // Only lat and lng
  feeling: string;
  request_time: string; // ISO 8601 timestamp
  user_id: string;
}

/**
 * Response for GET /break/plans
 */
export type BreakPlanResponse = BreakPlan[];

/**
 * Response for POST /break/plans
 * Returns multiple plans generated by LLM agent
 */
export type CreateBreakPlanResponse = BreakPlan[];

/**
 * Response for PATCH /break/plans/:id/chosen
 * Returns the updated plan
 */
export type UpdateBreakPlanChosenResponse = BreakPlan;

/**
 * Simplified history item for GET /history
 */
export interface HistoryItem {
  time: string; // ISO 8601 timestamp
  title: string;
  duration: number; // minutes
}

/**
 * Paginated history response with cursor-based pagination
 */
export interface HistoryResponse {
  items: HistoryItem[];
  cursor?: string; // ISO 8601 timestamp or UUID for next page
  hasMore: boolean;
}