# services/api/src/routes/break.ts
import type { FastifyInstance, FastifyRequest, FastifyReply } from 'fastify';
import fp from 'fastify-plugin';
import type {
  BreakPlanResponse,
  CreateBreakPlanRequest,
  CreateBreakPlanResponse,
  UpdateBreakPlanChosenResponse
} from '@take-a-break/types/break.js';
import type { SuccessResponse } from '@take-a-break/types/error.js';

interface BreakPlanParams {
  id: string;
}

interface AuthenticatedRequest extends FastifyRequest {
  user?: {
    uid: string;
    email?: string;
  };
}

async function breakRoutes(fastify: FastifyInstance) {
  /**
   * GET /break/plans
   * Get all break plans for the authenticated user
   */
  fastify.get<{ Reply: SuccessResponse<BreakPlanResponse> }>(
    '/break/plans',
    async (request: AuthenticatedRequest, reply: FastifyReply): Promise<SuccessResponse<BreakPlanResponse>> => {
      // TODO: Extract user_id from authenticated request
      const userId = request.user?.uid || '';

      // TODO: Fetch break plans from Firestore
      // Filter by user_id
      // Order by created_at descending

      const plans: BreakPlanResponse = [];

      return {
        success: true,
        data: plans
      };
    }
  );

  /**
   * POST /break/plans
   * Create multiple break plans (generated by LLM agent)
   */
  fastify.post<{
    Body: CreateBreakPlanRequest;
    Reply: SuccessResponse<CreateBreakPlanResponse>;
  }>(
    '/break/plans',
    async (
      request: FastifyRequest<{ Body: CreateBreakPlanRequest }>,
      reply: FastifyReply
    ): Promise<SuccessResponse<CreateBreakPlanResponse>> => {
      const body = request.body;

      // TODO: Extract user_id from authenticated request (JWT token)
      const userId = '';

      // TODO: Validate request body
      // - Check duration is positive
      // - Check location coordinates are valid
      // - Check feeling is not empty
      // - Check request_time is valid ISO 8601

      // TODO: Rate limiting check
      // - Check if user has exceeded 10 requests/hour limit

      // TODO: Reverse geocoding
      // - Convert coordinates to address using geocoding service
      const address = '';

      // TODO: Call LLM agent to generate break plans
      // - Input: duration, location (with address), feeling
      // - Output: Array of break plan suggestions with title, description, type

      // TODO: Create break plans in Firestore
      // - Generate UUID for each plan
      // - Set created_at and updated_at timestamps
      // - Set chosen: false
      // - Set order: 0, 1, 2, ...
      // - Set user_id from JWT token
      // - Store all generated plans

      const createdPlans: CreateBreakPlanResponse = [];

      reply.code(201);
      return {
        success: true,
        data: createdPlans
      };
    }
  );

  /**
   * PATCH /break/plans/:id/chosen
   * Mark a break plan as chosen
   */
  fastify.patch<{
    Params: BreakPlanParams;
    Reply: SuccessResponse<UpdateBreakPlanChosenResponse>;
  }>(
    '/break/plans/:id/chosen',
    async (
      request: FastifyRequest<{ Params: BreakPlanParams }>,
      reply: FastifyReply
    ): Promise<SuccessResponse<UpdateBreakPlanChosenResponse>> => {
      const { id } = request.params;

      // TODO: Extract user_id from authenticated request
      const userId = '';

      // TODO: Validate plan ID format (UUID)

      // TODO: Fetch break plan from Firestore
      // - Check if plan exists
      // - Check if plan belongs to user (authorization)

      // TODO: Update break plan in Firestore
      // - Set chosen: true
      // - Update updated_at timestamp

      const updatedPlan: UpdateBreakPlanChosenResponse = {
        id: '',
        title: '',
        description: '',
        duration: 0,
        location: {
          address: '',
          coordinates: { lat: 0, lng: 0 }
        },
        type: 'walk',
        feeling: '',
        chosen: true,
        order: 0,
        created_at: '',
        updated_at: '',
        user_id: ''
      };

      return {
        success: true,
        data: updatedPlan
      };
    }
  );
}

export default fp(breakRoutes, {
  name: 'break-routes'
});